Pharmacy Management System using Java (done during 4th sem)
This is my 4th sem project - Pharmacy Billing System using Java. It has been made using Java collection and frameworks and swings concepts to make an interactive GUI for the pharmacy billing system. It has features such as adding customer details, adding the medicine, updating the quantity, deleting the medicine and displaying the bill in a user effective way. It also has features where the user enters his/her age and depending on the age, the medicine will either be a syrup or a tablet.