C project done in 3rd sem
Graph analysis based on node degree is one of the most basic problem statement
arising in the migration of data servers to cloud. In this project, we first read through
a given data file and create the respective edges for each node. Each node is then
linked according to the data file and an adjacency file is created. By iterating through
the adjacency list, we create a graph structure. Each time a new source node is read,
a new graph object is created and the corresponding nodes to it are attached.
The graph structure is implemented and the degree of each source node is counted
and reported.